+++ CHANGELOG.md
@@ -9,5 +9,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Adds decorator `depends.dependable`
-* Adds class `depends.Depends`
+* Adds decorator `dependable.dependant`
+* Adds class `dependable.Depends`